Relocate HttpRetryConfig class to Reliability namespace
Summary: This commit relocates the HttpRetryConfig class, which defines the retry policy for HTTP requests, from the Configuration namespace to the Reliability namespace. This change improves the code organization and cohesion, as the class is more logically related to the DefaultHttpRetryHandler and its factory, which implement the retry logic. The commit also removes some unused code and simplifies some exception messages. The references and tests for the class are updated accordingly.
