+++ dotnet/src/SemanticKernel.UnitTests/Configuration/KernelConfigTests.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
 using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.SemanticKernel;
 using Microsoft.SemanticKernel.AI.OpenAI.Services;
 using Microsoft.SemanticKernel.Configuration;
@@ -66,69 +65,6 @@ public class KernelConfigTests
         // Act
         // Assert
         Assert.IsType<DefaultHttpRetryHandlerFactory>(config.HttpHandlerFactory);
-    }
-
-    [Fact]
-    public async Task NegativeMaxRetryCountThrowsAsync()
-    {
-        // Act
-        await Assert.ThrowsAsync<System.ArgumentOutOfRangeException>(() =>
-        {
-            var httpRetryConfig = new KernelConfig.HttpRetryConfig() { MaxRetryCount = -1 };
-            return Task.CompletedTask;
-        });
-    }
-
-    [Fact]
-    public void SetDefaultHttpRetryConfig()
-    {
-        // Arrange
-        var config = new KernelConfig();
-        var httpRetryConfig = new KernelConfig.HttpRetryConfig() { MaxRetryCount = 1 };
-
-        // Act
-        config.SetDefaultHttpRetryConfig(httpRetryConfig);
-
-        // Assert
-        Assert.Equal(httpRetryConfig, config.DefaultHttpRetryConfig);
-    }
-
-    [Fact]
-    public void SetDefaultHttpRetryConfigToDefaultIfNotSet()
-    {
-        // Arrange
-        var config = new KernelConfig();
-
-        // Act
-        // Assert
-        var defaultConfig = new KernelConfig.HttpRetryConfig();
-        Assert.Equal(defaultConfig.MaxRetryCount, config.DefaultHttpRetryConfig.MaxRetryCount);
-        Assert.Equal(defaultConfig.MaxRetryDelay, config.DefaultHttpRetryConfig.MaxRetryDelay);
-        Assert.Equal(defaultConfig.MinRetryDelay, config.DefaultHttpRetryConfig.MinRetryDelay);
-        Assert.Equal(defaultConfig.MaxTotalRetryTime, config.DefaultHttpRetryConfig.MaxTotalRetryTime);
-        Assert.Equal(defaultConfig.UseExponentialBackoff, config.DefaultHttpRetryConfig.UseExponentialBackoff);
-        Assert.Equal(defaultConfig.RetryableStatusCodes, config.DefaultHttpRetryConfig.RetryableStatusCodes);
-        Assert.Equal(defaultConfig.RetryableExceptionTypes, config.DefaultHttpRetryConfig.RetryableExceptionTypes);
-    }
-
-    [Fact]
-    public void SetDefaultHttpRetryConfigToDefaultIfNull()
-    {
-        // Arrange
-        var config = new KernelConfig();
-
-        // Act
-        config.SetDefaultHttpRetryConfig(null);
-
-        // Assert
-        var defaultConfig = new KernelConfig.HttpRetryConfig();
-        Assert.Equal(defaultConfig.MaxRetryCount, config.DefaultHttpRetryConfig.MaxRetryCount);
-        Assert.Equal(defaultConfig.MaxRetryDelay, config.DefaultHttpRetryConfig.MaxRetryDelay);
-        Assert.Equal(defaultConfig.MinRetryDelay, config.DefaultHttpRetryConfig.MinRetryDelay);
-        Assert.Equal(defaultConfig.MaxTotalRetryTime, config.DefaultHttpRetryConfig.MaxTotalRetryTime);
-        Assert.Equal(defaultConfig.UseExponentialBackoff, config.DefaultHttpRetryConfig.UseExponentialBackoff);
-        Assert.Equal(defaultConfig.RetryableStatusCodes, config.DefaultHttpRetryConfig.RetryableStatusCodes);
-        Assert.Equal(defaultConfig.RetryableExceptionTypes, config.DefaultHttpRetryConfig.RetryableExceptionTypes);
     }
 
     [Fact]
